Rename 'features' folder to '_contents'
