+++ apps/nextjs-website/src/_contents/appIO/appIO.ts
@@ -1,5 +1,5 @@
 import { Product } from '@/lib/types/product';
-import { appIoOverviewPath } from '@/features/appIO/overviewPath';
+import { appIoOverviewPath } from '@/_contents/appIO/overviewPath';
 
 export const appIO: Product = {
   name: 'App IO',
